Package b0 0.0.1

WARNING. This package is unstable and work in progress, do not depend on it.

B0 describes software construction and deployments using modular and customizable definitions written in OCaml.

It also provides the B00 build library which provides arbitrary build abstraction with reliable, efficient incremental rebuilds.

Manuals

For B0 and B00 users.

The B0 manual is a conceptual overview of the B0 system and B0 files. It is recommended reading.

The B00 manual describes the B00 build library.

For B0 programmers.

The B0 driver development manual describes how to make your own drivers to operate on B0 files.

The TODO page

Old manual fragments and TODO which will eventually disapear.

B00

B00 is the build model used by B0. It can be used for its own good to devise domain specific build system – see for example the brzo or odig tools.

The b0.b00.std and b0.b00.kit libraries also provide sharp tools for quick "scripts" and command line tools even if you are not using b0.b00.

Library b0.b00.std

The b0.b00.std library has a few things that should be in the stdlib and a few others that should not be but are useful for B0 and B00 based programs.

B00_std Standard library needs.

Library b0.b00

The build library.

B000 Build kernel plumbing.

B000_conv Serialize and format B000 values.

B00 Build kernel

Library b0.b00.kit

Toolkit for working with B00 and B0. A few things do not depend on the b0.b00 library but are useful to have under your hand.

IO

B00_base64 Base64 codec.

B00_fexts Sets of file extension sets.

B00_findex File indexes.

B00_htmlg HTML generation.

B00_http World Wide Web interaction.

B00_lines Parse text lines.

B00_serialk_json JSON text support.

B00_serialk_sexp S-expression support.

B00_serialk_text UTF-8 text lexing tools.

Dev tools and info

B00_cmark B00 cmark support.

B00_github GitHub interaction.

B00_ocaml B0 ocaml support.

B00_odoc B0 odoc support.

B00_os Operating system and machine information.

B00_vcs Version control system (VCS) repositories.

UI fragments

Mostly Cmdliner user interface fragments.

B00_ui User interaction fragments.

B00_std_ui B00_std setup and cli fragments.

B00_trace Trace build operations.

Viewers and editors

B00_editor Editor interaction.

B00_pager Pager interaction.

B00_pdf_viewer PDF viewer interaction.

B00_www_browser Web browser interaction.

B0

B0 is the system for describing software construction and deployments via B0 files.

Library b0

B0_def B0 definitions.

B0_meta Metadata.

B0_build B0 builds.

B0_unit Build units.

B0_pack Build packs.

Library b0.kit

Toolkit for B0 files.

B0_kit APIs to use in B0 files.

B0_srcs Select source files.

B0_ocaml OCaml B0 file support.

Library b0.driver

The driver library provides programmatic access to the definitions of B0 files. Only needed if you want to process B0 files your own own way.

B0_file B0 file source and expansion.

B0_driver B0 file drivers.

Package info

authors

The b0 programmers

changes-files

CHANGES.md

depends

cmdliner

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/b0

issues

https://github.com/b0-system/b0/issues

license

BSD2

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/b0/doc

readme-files

README.md

repo

git+https://erratique.ch/repos/b0.git#b00

tags

build

dev

org:b0-system

org:erratique

version

0.0.1